+++ src/main/kotlin/crimera/patches/twitter/interaction/downloads/changedirectory/fingerprints/SetDownloadDestinationFingerprint.kt
@@ -1,0 +1,10 @@
+package crimera.patches.twitter.interaction.downloads.changedirectory.fingerprints
+
+import app.revanced.patcher.fingerprint.MethodFingerprint
+
+object SetDownloadDestinationFingerprint: MethodFingerprint(
+    returnType = "V",
+    strings = listOf(
+        "parse(downloadData.url)"
+    )
+)