+++ src/main/kotlin/crimera/patches/twitter/misc/settings/fingerprints/SettingsFingerprint.kt
@@ -1,0 +1,13 @@
+package crimera.patches.twitter.misc.settings.fingerprints
+
+import app.revanced.patcher.fingerprint.MethodFingerprint
+
+object SettingsFingerprint: MethodFingerprint(
+    returnType = "V",
+    strings = listOf(
+        "pref_proxy"
+    ),
+    customFingerprint = { it, _ ->
+        it.name == "<clinit>"
+    }
+)