+++ src/main/kotlin/crimera/patches/twitter/misc/viewcount/RemoveViewCountPatch.kt
@@ -1,4 +1,4 @@
-package crimera.patches.twitter.view
+package crimera.patches.twitter.misc.viewcount
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
@@ -8,7 +8,7 @@ import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
-import crimera.patches.twitter.view.fingerprints.RemoveViewCountPatchFingerprint
+import crimera.patches.twitter.misc.viewcount.fingerprints.RemoveViewCountPatchFingerprint
 
 // Credits to @iKirby
 @Patch(